# Backend/Syslab/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "pablokauan.tech@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/pessoas": {
            "get": {
                "tags": [
                    "Pessoa"
                ],
                "summary": "Listar pessoas",
                "operationId": "7c7d2c9decc48109bcd470035e629ecb",
                "responses": {
                    "200": {
                        "description": "Lista de pessoas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Pessoa"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pessoa"
                ],
                "summary": "Criar pessoa",
                "operationId": "b72258a0649966eb7cb39a9fdc67bf4b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "research": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "lattes": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "orcid": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "cargo": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "curso": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pessoa criada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pessoa"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/pessoas/{id}": {
            "get": {
                "tags": [
                    "Pessoa"
                ],
                "summary": "Buscar pessoa",
                "operationId": "9767860b5875b66fc3535a00d47aec2d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pessoa encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pessoa"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pessoa não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pessoa"
                ],
                "summary": "Atualizar pessoa",
                "operationId": "da7d1255c980a411d322994cbca98632",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "research": {
                                        "type": "string"
                                    },
                                    "lattes": {
                                        "type": "string"
                                    },
                                    "orcid": {
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "type": "string"
                                    },
                                    "cargo": {
                                        "type": "string"
                                    },
                                    "curso": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pessoa atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pessoa"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pessoa"
                ],
                "summary": "Remover pessoa",
                "operationId": "6ac497549c4c86a2d7bbd144d43dffb3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removido"
                    }
                }
            }
        },
        "/api/pessoas/{id}/projetos": {
            "get": {
                "tags": [
                    "Pessoa-Projetos"
                ],
                "summary": "Listar projetos da pessoa",
                "operationId": "e3cd91b744bdc63c549cb6140d99eda1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Projetos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Projeto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pessoa-Projetos"
                ],
                "summary": "Vincular projeto",
                "operationId": "3acbd4c60b65fc5bea1d79fa0a56718f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "projeto_id"
                                ],
                                "properties": {
                                    "projeto_id": {
                                        "type": "integer"
                                    },
                                    "integrantes": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Projeto vinculado"
                    }
                }
            }
        },
        "/api/pessoas/{id}/projetos/sync": {
            "put": {
                "tags": [
                    "Pessoa-Projetos"
                ],
                "summary": "Sincronizar projetos",
                "operationId": "392d9939df0e806400786053ab79dbf6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "projetos": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Projetos sincronizados"
                    }
                }
            }
        },
        "/api/pessoas/{id}/projetos/{projetoId}": {
            "delete": {
                "tags": [
                    "Pessoa-Projetos"
                ],
                "summary": "Desvincular projeto",
                "operationId": "690b9db118d27d6aef0492f951717f89",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "projetoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Projeto removido"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "News": {
                "title": "News",
                "description": "Notícia publicada no sistema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Novo laboratório é inaugurado"
                    },
                    "subtitle": {
                        "type": "string",
                        "example": "Espaço amplia pesquisas na universidade",
                        "nullable": true
                    },
                    "body_text": {
                        "type": "string",
                        "example": "O laboratório foi inaugurado com foco em inovação..."
                    },
                    "image": {
                        "type": "string",
                        "example": "news.jpg",
                        "nullable": true
                    },
                    "datastamp": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-01T10:30:00Z"
                    },
                    "author_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "author": {
                        "$ref": "#/components/schemas/Pessoa"
                    }
                },
                "type": "object"
            },
            "Pessoa": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Pablo Kauan"
                    },
                    "image": {
                        "type": "string",
                        "example": "perfil.jpg",
                        "nullable": true
                    },
                    "research": {
                        "type": "string",
                        "example": "Inteligência Artificial",
                        "nullable": true
                    },
                    "lattes": {
                        "type": "string",
                        "nullable": true
                    },
                    "orcid": {
                        "type": "string",
                        "nullable": true
                    },
                    "descricao": {
                        "type": "string",
                        "nullable": true
                    },
                    "cargo": {
                        "type": "string",
                        "nullable": true
                    },
                    "curso": {
                        "type": "string",
                        "nullable": true
                    },
                    "projetos": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Projeto"
                        }
                    },
                    "news": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/News"
                        }
                    }
                },
                "type": "object"
            },
            "Projeto": {
                "title": "Projeto",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Sistema de Monitoramento Energético"
                    },
                    "image": {
                        "type": "string",
                        "example": "projeto.jpg",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Projeto voltado ao monitoramento de consumo energético",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "ativo"
                    },
                    "date_begin": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-01-01"
                    },
                    "date_end": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-31",
                        "nullable": true
                    },
                    "pivot": {
                        "properties": {
                            "integrantes": {
                                "type": "string",
                                "example": "Desenvolvedor Backend",
                                "nullable": true
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "pessoas": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Pessoa"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}